+++ src/index.ts
@@ -12,5 +12,6 @@ export * from './core/models/events';
 export * from './core/models/setup';
 export * from './core/models/utils';
 export * from './app/initializer'
+export { CuiInstance } from './app/instance';
 
 window.cuiInit = new CuiInit();