+++ changelog.md
@@ -1,7 +1,12 @@
 # bpd cUI Light changelog
-# [0.1.9] WIP
+
+# [0.1.10] WIP
+# [0.1.9] 2020-08-08
+* [Fixed] Export CuiInstance class
+* [Change] Add mutation observer to accordion to detect children changes
+
 # [0.1.8] 2020-08-06
-* [Fix] Offcanvas background
+* [Fixed] Offcanvas background
 # [0.1.7] 2020-08-06
 * [Fixed] Dialog and offcanvas background when opened
 # [0.1.6] 2020-08-06